feat: enhance purchase order functionality with product-based retrieval and data type updates

+++ server/src/main/java/com/express/inventory/api/purchases/PurchaseOrder.java
@@ -31,6 +31,6 @@ public class PurchaseOrder {
     private String orderStatus;
     private BigDecimal orderPrice;
     private String purchaseDate;
-    private Double totalQuantity;
+    private BigDecimal totalQuantity;
 
 }
